requirements.txt
# Core Setup
setuptools>=41.0.0

# UI Framework
customtkinter==5.2.0
darkdetect==0.7.1  # Required for customtkinter theme detection

# Audio/Video Processing
ffmpeg-python==0.2.0
spleeter==2.3.0
soundfile==0.11.0
pydub==0.25.1
librosa==0.8.0

# Machine Learning & Scientific Computing
tensorflow==2.5.0
numpy==1.19.5
scipy==1.7.0
scikit-learn==0.24.2
pandas==1.1.5

# ML Dependencies (required by TensorFlow & Spleeter)
numba==0.53.1
llvmlite==0.36.0
grpcio==1.34.1
tensorboard==2.5.0
keras-preprocessing==1.1.2
h5py==3.1.0
protobuf==3.19.6
google-pasta==0.2.0
opt-einsum==3.3.0
tensorflow-estimator==2.5.0
google-auth==1.35.0

# Utilities & Compatibility
six==1.15.0
typing-extensions>=3.7.4 
wheel==0.36.2
wrapt==1.12.1
absl-py==0.12.0
markdown==3.3.4
werkzeug==2.0.3
platformdirs>=4.0.0  